requirements.txt
asgiref==3.4.1
astroid==2.9.0
black==21.12b0
click==8.0.3
Django==3.2
django-debug-toolbar==2.2
django-split-settings==1.1.0
flake8==4.0.1
gunicorn==20.0.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
wrapt==1.13.3
